client: Add a basic module to interact with HAProxy
The `client.go` module, implements the package's main API with HAProxy's
UNIX Socket interface.

It consists of the `HAProxyClient` type that is the main component of
the current module and corresponds to a HAProxy connection object. In
order to send a command to HAProxy you should use the `SendCommand`
function. The command argument should be in string format. As for the
return values, we currently don't perform any kind of sanity check in
the server's response; we just convert it to string and return it to the
caller.

Finally, note that currently you can only interact with HAProxy via a
UNIX socket. An upcoming patch will provide support for listening to a
TCP socket, too.