client: Allow connections of tcp4 and tcp6 types
Until now, we could only interract with HAProxy via a UNIX domain
socket. Since HAProxy can also listen to either a v4 or v6 TCP port,
we've included that functionality for our client, too.

Note that based on the network type, the respective validity function
will be called on the `Address` argument. However, a probable error will
be only raised on the first attempt to send a command to HAProxy due to
the lazy nature a connection to HAProxy is created. This approach will
be modified in the future in order to be directly informed about a wrong
address format without having to wait a command be send to the server.