tests: Setup the testing structure template
In this commit, we setup the main structure of the client's testing
functionality.

The test suite of the go-haproxy client is located under a separate
directory, named `tests/`, on the contrary to the official Golang way of
defining tests. The main reason behind this approach is simple: I don't
like maintaining my test files along with the main source code.

However, this approach causes an obvious issue, i.e., you can not import
and subsequently test the client's non-exported functions. A temporary
workaround in order to be able testing all client's functions and also
continue maintaing the test suite in a separate folder, is to symlink
the complete test suite to its proper location prior to running it, and
unlink it as soon as the test run completes, either successfully or not.

Also, we've introduced a new testing module named `helpers_test.go`,
which will contain any helper/utility functionality usefull to the
package's test suite. Currently, we include three helper functions,
which are the following:

- assertNil
- assertEqual
- assertNotEqual